Fix: split comparison selector support
